Fixed the code repetition, fixed run error with March
