+++ Code/model.py
@@ -1,5 +1,5 @@
 """
-An initial XGBoost model based on 
+An initial XGBoost model based on
 https://github.com/dmlc/xgboost/blob/master/demo/guide-python/multioutput_regression.py
 
 """
@@ -11,7 +11,7 @@ from sklearn.model_selection import train_test_split
 
 def rmseModel():
     """ Root mean squared error XGBoost trained on the june data"""
-    X, y = loadModelData(month="June",threshold=-86, verbose=False)
+    X, y = loadModelData(month="June",threshold=-102, verbose=False)
 
     X_train, X_remaining, y_train, y_remaining = train_test_split(X, y, train_size=0.8, random_state=101)
     X_valid, X_test, y_valid, y_test = train_test_split(X_remaining,y_remaining, test_size=0.5)
