+++ Code/plot.py
@@ -1,5 +1,5 @@
 from matplotlib import pyplot as plt
-from utils import loadData, loadNodes, loadBoundaries
+from utils import loadData, loadNodes, loadBoundaries, deriveEquation
 from multilat import marchPredictions, junePredictions
 import os, argparse
 
@@ -77,7 +77,28 @@ def plotOriginal(month="March", threshold=-90):
         plt.savefig(path+"/"+str(test)+".png", bbox_inches='tight')
         # bbox_inches removes extra white spaces
         plt.clf()
+def plotEquation():
+    # 100 linearly spaced numbers
+    x = np.linspace(0,200,10000)
+    values = utils.deriveEquation()
+    print(values)
+    y = values[0]*np.exp(x*values[1])+values[2]
 
+    # setting the axes at the centre
+    fig = plt.figure()
+    ax = fig.add_subplot(1, 1, 1)
+    ax.spines['left'].set_position('center')
+    ax.spines['bottom'].set_position('zero')
+    ax.spines['right'].set_color('none')
+    ax.spines['top'].set_color('none')
+    ax.xaxis.set_ticks_position('bottom')
+    ax.yaxis.set_ticks_position('left')
+
+    # plot the function
+    plt.plot(x,y, 'r')
+
+    # show the plot
+    plt.show()
 
 def main(args=None):
     rssiThreshold=-105.16
